chore: address all PR feedback, and address a bug in which we accidentally expected a SIP-003 length prefix in chunks that we parsed
