+++ testnet/stacks-node/src/tests/mod.rs
@@ -48,6 +48,7 @@ mod epoch_24;
 mod integrations;
 mod mempool;
 pub mod neon_integrations;
+mod stackerdb;
 
 // $ cat /tmp/out.clar
 pub const STORE_CONTRACT: &str = r#"(define-map store { key: (string-ascii 32) } { value: (string-ascii 32) })
