+++ testnet/stacks-node/src/tests/neon_integrations.rs
@@ -28,7 +28,6 @@ use stacks::core::{
     BLOCK_LIMIT_MAINNET_21, CHAIN_ID_TESTNET, HELIUM_BLOCK_LIMIT_20, PEER_VERSION_EPOCH_1_0,
     PEER_VERSION_EPOCH_2_0, PEER_VERSION_EPOCH_2_05, PEER_VERSION_EPOCH_2_1,
 };
-use stacks::libstackerdb::{StackerDBChunkAckData, StackerDBChunkData};
 use stacks::net::atlas::{AtlasConfig, AtlasDB, MAX_ATTACHMENT_INV_PAGES_PER_REQUEST};
 use stacks::net::{
     AccountEntryResponse, ContractSrcResponse, GetAttachmentResponse, GetAttachmentsInvResponse,
@@ -98,7 +97,6 @@ use stacks::chainstate::stacks::miner::{
     TransactionSuccessEvent,
 };
 use stacks::net::RPCFeeEstimateResponse;
-use stacks::vm::types::QualifiedContractIdentifier;
 use stacks::vm::ClarityName;
 use stacks::vm::ContractName;
 use std::convert::TryFrom;
@@ -10836,212 +10834,3 @@ fn microblock_miner_multiple_attempts() {
     channel.stop_chains_coordinator();
 }
 
-fn post_stackerdb_chunk(
-    http_origin: &str,
-    stackerdb_contract_id: &QualifiedContractIdentifier,
-    data: Vec<u8>,
-    signer: &StacksPrivateKey,
-    slot_id: u32,
-    slot_version: u32,
-) -> StackerDBChunkAckData {
-    let mut chunk = StackerDBChunkData::new(slot_id, slot_version, data);
-    chunk.sign(&signer).unwrap();
-
-    let chunk_body = serde_json::to_string(&chunk).unwrap();
-
-    let client = reqwest::blocking::Client::new();
-    let path = format!(
-        "{}/v2/stackerdb/{}/{}/chunks",
-        http_origin,
-        &StacksAddress::from(stackerdb_contract_id.issuer.clone()),
-        stackerdb_contract_id.name
-    );
-    let res = client
-        .post(&path)
-        .header("Content-Type", "application/json")
-        .body(chunk_body.as_bytes().to_vec())
-        .send()
-        .unwrap();
-    if res.status().is_success() {
-        let ack: StackerDBChunkAckData = res.json().unwrap();
-        info!("Got stackerdb ack: {:?}", &ack);
-        return ack;
-    } else {
-        eprintln!("StackerDB post error: {}", res.text().unwrap());
-        panic!("");
-    }
-}
-
-fn get_stackerdb_chunk(
-    http_origin: &str,
-    stackerdb_contract_id: &QualifiedContractIdentifier,
-    slot_id: u32,
-    slot_version: Option<u32>,
-) -> Vec<u8> {
-    let path = if let Some(version) = slot_version {
-        format!(
-            "{}/v2/stackerdb/{}/{}/{}/{}",
-            http_origin,
-            StacksAddress::from(stackerdb_contract_id.issuer.clone()),
-            stackerdb_contract_id.name,
-            slot_id,
-            version
-        )
-    } else {
-        format!(
-            "{}/v2/stackerdb/{}/{}/{}",
-            http_origin,
-            StacksAddress::from(stackerdb_contract_id.issuer.clone()),
-            stackerdb_contract_id.name,
-            slot_id
-        )
-    };
-
-    let client = reqwest::blocking::Client::new();
-    let res = client.get(&path).send().unwrap();
-
-    if res.status().is_success() {
-        let chunk_data: Vec<u8> = res.bytes().unwrap().to_vec();
-        return chunk_data;
-    } else {
-        eprintln!("Get chunk error: {}", res.text().unwrap());
-        panic!("");
-    }
-}
-
-#[test]
-#[ignore]
-fn test_stackerdb_load_store() {
-    if env::var("BITCOIND_TEST") != Ok("1".into()) {
-        return;
-    }
-
-    let (mut conf, _) = neon_integration_test_conf();
-    conf.events_observers.push(EventObserverConfig {
-        endpoint: format!("localhost:{}", test_observer::EVENT_OBSERVER_PORT),
-        events_keys: vec![EventKeyType::AnyEvent],
-    });
-
-    let privks = vec![
-        // ST2DS4MSWSGJ3W9FBC6BVT0Y92S345HY8N3T6AV7R
-        StacksPrivateKey::from_hex(
-            "9f1f85a512a96a244e4c0d762788500687feb97481639572e3bffbd6860e6ab001",
-        )
-        .unwrap(),
-        // STVN97YYA10MY5F6KQJHKNYJNM24C4A1AT39WRW
-        StacksPrivateKey::from_hex(
-            "94c319327cc5cd04da7147d32d836eb2e4c44f4db39aa5ede7314a761183d0c701",
-        )
-        .unwrap(),
-    ];
-
-    let stackerdb_contract = "
-        ;; stacker DB
-        (define-read-only (stackerdb-get-signer-slots)
-            (ok (list
-                {
-                    signer: 'ST2DS4MSWSGJ3W9FBC6BVT0Y92S345HY8N3T6AV7R,
-                    num-slots: u3
-                }
-                {
-                    signer: 'STVN97YYA10MY5F6KQJHKNYJNM24C4A1AT39WRW,
-                    num-slots: u3
-                })))
-
-        (define-read-only (stackerdb-get-config)
-            (ok {
-                chunk-size: u4096,
-                write-freq: u0,
-                max-writes: u4096,
-                max-neighbors: u32,
-                hint-replicas: (list )
-            }))
-    ";
-
-    conf.initial_balances.append(&mut vec![
-        InitialBalance {
-            address: to_addr(&privks[0]).into(),
-            amount: 10_000_000_000_000,
-        },
-        InitialBalance {
-            address: to_addr(&privks[1]).into(),
-            amount: 10_000_000_000_000,
-        },
-    ]);
-
-    conf.node.stacker_dbs.push(QualifiedContractIdentifier::new(
-        to_addr(&privks[0]).into(),
-        "hello-world".into(),
-    ));
-    let contract_id = conf.node.stacker_dbs[0].clone();
-
-    test_observer::spawn();
-
-    let mut btcd_controller = BitcoinCoreController::new(conf.clone());
-    btcd_controller
-        .start_bitcoind()
-        .map_err(|_e| ())
-        .expect("Failed starting bitcoind");
-
-    let mut btc_regtest_controller = BitcoinRegtestController::new(conf.clone(), None);
-
-    btc_regtest_controller.bootstrap_chain(201);
-
-    eprintln!("Chain bootstrapped...");
-
-    let mut run_loop = neon::RunLoop::new(conf.clone());
-    let blocks_processed = run_loop.get_blocks_processed_arc();
-
-    thread::spawn(move || run_loop.start(None, 0));
-
-    // Give the run loop some time to start up!
-    eprintln!("Wait for runloop...");
-    wait_for_runloop(&blocks_processed);
-
-    // First block wakes up the run loop.
-    eprintln!("Mine first block...");
-    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
-
-    // Second block will hold our VRF registration.
-    eprintln!("Mine second block...");
-    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
-
-    // Third block will be the first mined Stacks block.
-    eprintln!("Mine third block...");
-    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
-
-    let http_origin = format!("http://{}", &conf.node.rpc_bind);
-
-    eprintln!("Send contract-publish...");
-    let tx = make_contract_publish(&privks[0], 0, 10_000, "hello-world", stackerdb_contract);
-    submit_tx(&http_origin, &tx);
-
-    // mine it
-    eprintln!("Mine it...");
-    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
-    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
-
-    // write some chunks and read them back
-    for i in 0..3 {
-        let chunk_str = format!("Hello chunks {}", &i);
-        let ack = post_stackerdb_chunk(
-            &http_origin,
-            &contract_id,
-            chunk_str.as_bytes().to_vec(),
-            &privks[0],
-            0,
-            (i + 1) as u32,
-        );
-        debug!("ACK: {:?}", &ack);
-
-        let data = get_stackerdb_chunk(&http_origin, &contract_id, 0, Some((i + 1) as u32));
-        assert_eq!(data, chunk_str.as_bytes().to_vec());
-
-        let data = get_stackerdb_chunk(&http_origin, &contract_id, 0, None);
-        assert_eq!(data, chunk_str.as_bytes().to_vec());
-    }
-
-    loop {
-        sleep_ms(1000);
-    }
-}
