+++ testnet/stacks-node/src/tests/stackerdb.rs
@@ -1,0 +1,261 @@
+// Copyright (C) 2013-2020 Blockstack PBC, a public benefit corporation
+// Copyright (C) 2020-2023 Stacks Open Internet Foundation
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+use std::env;
+use std::thread;
+
+use super::bitcoin_regtest::BitcoinCoreController;
+use crate::{
+    burnchains::BurnchainController,
+    config::EventKeyType,
+    config::EventObserverConfig,
+    config::InitialBalance,
+    neon,
+    BitcoinRegtestController,
+    tests::{
+        to_addr,
+        make_contract_publish,
+        neon_integrations::{
+            neon_integration_test_conf,
+            test_observer,
+            wait_for_runloop,
+            next_block_and_wait,
+            submit_tx
+        }
+    },
+};
+
+use stacks::{
+    chainstate::stacks::{
+        StacksPrivateKey,
+    },
+};
+
+use clarity::vm::types::QualifiedContractIdentifier;
+
+use stacks::libstackerdb::{StackerDBChunkData, StackerDBChunkAckData};
+
+use stacks_common::types::chainstate::StacksAddress;
+
+use serde_json;
+
+use reqwest;
+
+fn post_stackerdb_chunk(
+    http_origin: &str,
+    stackerdb_contract_id: &QualifiedContractIdentifier,
+    data: Vec<u8>,
+    signer: &StacksPrivateKey,
+    slot_id: u32,
+    slot_version: u32,
+) -> StackerDBChunkAckData {
+    let mut chunk = StackerDBChunkData::new(slot_id, slot_version, data);
+    chunk.sign(&signer).unwrap();
+
+    let chunk_body = serde_json::to_string(&chunk).unwrap();
+
+    let client = reqwest::blocking::Client::new();
+    let path = format!(
+        "{}/v2/stackerdb/{}/{}/chunks",
+        http_origin,
+        &StacksAddress::from(stackerdb_contract_id.issuer.clone()),
+        stackerdb_contract_id.name
+    );
+    let res = client
+        .post(&path)
+        .header("Content-Type", "application/json")
+        .body(chunk_body.as_bytes().to_vec())
+        .send()
+        .unwrap();
+    if res.status().is_success() {
+        let ack: StackerDBChunkAckData = res.json().unwrap();
+        info!("Got stackerdb ack: {:?}", &ack);
+        return ack;
+    } else {
+        eprintln!("StackerDB post error: {}", res.text().unwrap());
+        panic!("");
+    }
+}
+
+fn get_stackerdb_chunk(
+    http_origin: &str,
+    stackerdb_contract_id: &QualifiedContractIdentifier,
+    slot_id: u32,
+    slot_version: Option<u32>,
+) -> Vec<u8> {
+    let path = if let Some(version) = slot_version {
+        format!(
+            "{}/v2/stackerdb/{}/{}/{}/{}",
+            http_origin,
+            StacksAddress::from(stackerdb_contract_id.issuer.clone()),
+            stackerdb_contract_id.name,
+            slot_id,
+            version
+        )
+    } else {
+        format!(
+            "{}/v2/stackerdb/{}/{}/{}",
+            http_origin,
+            StacksAddress::from(stackerdb_contract_id.issuer.clone()),
+            stackerdb_contract_id.name,
+            slot_id
+        )
+    };
+
+    let client = reqwest::blocking::Client::new();
+    let res = client.get(&path).send().unwrap();
+
+    if res.status().is_success() {
+        let chunk_data: Vec<u8> = res.bytes().unwrap().to_vec();
+        return chunk_data;
+    } else {
+        eprintln!("Get chunk error: {}", res.text().unwrap());
+        panic!("");
+    }
+}
+
+#[test]
+#[ignore]
+fn test_stackerdb_load_store() {
+    if env::var("BITCOIND_TEST") != Ok("1".into()) {
+        return;
+    }
+
+    let (mut conf, _) = neon_integration_test_conf();
+    conf.events_observers.push(EventObserverConfig {
+        endpoint: format!("localhost:{}", test_observer::EVENT_OBSERVER_PORT),
+        events_keys: vec![EventKeyType::AnyEvent],
+    });
+
+    let privks = vec![
+        // ST2DS4MSWSGJ3W9FBC6BVT0Y92S345HY8N3T6AV7R
+        StacksPrivateKey::from_hex(
+            "9f1f85a512a96a244e4c0d762788500687feb97481639572e3bffbd6860e6ab001",
+        )
+        .unwrap(),
+        // STVN97YYA10MY5F6KQJHKNYJNM24C4A1AT39WRW
+        StacksPrivateKey::from_hex(
+            "94c319327cc5cd04da7147d32d836eb2e4c44f4db39aa5ede7314a761183d0c701",
+        )
+        .unwrap(),
+    ];
+
+    let stackerdb_contract = "
+        ;; stacker DB
+        (define-read-only (stackerdb-get-signer-slots)
+            (ok (list
+                {
+                    signer: 'ST2DS4MSWSGJ3W9FBC6BVT0Y92S345HY8N3T6AV7R,
+                    num-slots: u3
+                }
+                {
+                    signer: 'STVN97YYA10MY5F6KQJHKNYJNM24C4A1AT39WRW,
+                    num-slots: u3
+                })))
+
+        (define-read-only (stackerdb-get-config)
+            (ok {
+                chunk-size: u4096,
+                write-freq: u0,
+                max-writes: u4096,
+                max-neighbors: u32,
+                hint-replicas: (list )
+            }))
+    ";
+
+    conf.initial_balances.append(&mut vec![
+        InitialBalance {
+            address: to_addr(&privks[0]).into(),
+            amount: 10_000_000_000_000,
+        },
+        InitialBalance {
+            address: to_addr(&privks[1]).into(),
+            amount: 10_000_000_000_000,
+        },
+    ]);
+
+    conf.node.stacker_dbs.push(QualifiedContractIdentifier::new(
+        to_addr(&privks[0]).into(),
+        "hello-world".into(),
+    ));
+    let contract_id = conf.node.stacker_dbs[0].clone();
+
+    test_observer::spawn();
+
+    let mut btcd_controller = BitcoinCoreController::new(conf.clone());
+    btcd_controller
+        .start_bitcoind()
+        .map_err(|_e| ())
+        .expect("Failed starting bitcoind");
+
+    let mut btc_regtest_controller = BitcoinRegtestController::new(conf.clone(), None);
+
+    btc_regtest_controller.bootstrap_chain(201);
+
+    eprintln!("Chain bootstrapped...");
+
+    let mut run_loop = neon::RunLoop::new(conf.clone());
+    let blocks_processed = run_loop.get_blocks_processed_arc();
+
+    thread::spawn(move || run_loop.start(None, 0));
+
+    // Give the run loop some time to start up!
+    eprintln!("Wait for runloop...");
+    wait_for_runloop(&blocks_processed);
+
+    // First block wakes up the run loop.
+    eprintln!("Mine first block...");
+    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
+
+    // Second block will hold our VRF registration.
+    eprintln!("Mine second block...");
+    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
+
+    // Third block will be the first mined Stacks block.
+    eprintln!("Mine third block...");
+    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
+
+    let http_origin = format!("http://{}", &conf.node.rpc_bind);
+
+    eprintln!("Send contract-publish...");
+    let tx = make_contract_publish(&privks[0], 0, 10_000, "hello-world", stackerdb_contract);
+    submit_tx(&http_origin, &tx);
+
+    // mine it
+    eprintln!("Mine it...");
+    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
+    next_block_and_wait(&mut btc_regtest_controller, &blocks_processed);
+
+    // write some chunks and read them back
+    for i in 0..3 {
+        let chunk_str = format!("Hello chunks {}", &i);
+        let ack = post_stackerdb_chunk(
+            &http_origin,
+            &contract_id,
+            chunk_str.as_bytes().to_vec(),
+            &privks[0],
+            0,
+            (i + 1) as u32,
+        );
+        debug!("ACK: {:?}", &ack);
+
+        let data = get_stackerdb_chunk(&http_origin, &contract_id, 0, Some((i + 1) as u32));
+        assert_eq!(data, chunk_str.as_bytes().to_vec());
+
+        let data = get_stackerdb_chunk(&http_origin, &contract_id, 0, None);
+        assert_eq!(data, chunk_str.as_bytes().to_vec());
+    }
+}
